+++ question/type/mtf/edit_mtf_form.php
@@ -530,29 +530,14 @@ class qtype_mtf_edit_form extends question_edit_form {
             }
         }
 
-        // Can only have one idnumber per category.
-        if (strpos($data['category'], ',') !== false) {
-            list($category, $categorycontextid) = explode(',', $data['category']);
-        } else {
-            $category = $data['category'];
-        }
 
+        // Can only have one idnumber.
         if (isset($data['idnumber']) && ((string)$data['idnumber'] !== '')) {
-            if (empty($data['usecurrentcat']) && !empty($data['categorymoveto'])) {
-                $categoryinfo = $data['categorymoveto'];
-            } else {
-                $categoryinfo = $data['category'];
-            }
-
-            list($categoryid, $notused) = explode(',', $categoryinfo);
-            $conditions = 'category = ? AND idnumber = ?';
-            $params = [$categoryid, $data['idnumber']];
-
+            $conditions = 'idnumber = ?';
             if (!empty($this->question->id)) {
                 $conditions .= ' AND id <> ?';
                 $params[] = $this->question->id;
             }
-
             if ($DB->record_exists_select('question', $conditions, $params)) {
                 $errors['idnumber'] = get_string('idnumbertaken', 'error');
             }