+++ theme/boost_union/lang/en/theme_boost_union.php
@@ -57,17 +57,6 @@ $string['faviconsetting'] = 'Favicon';
 $string['faviconsetting_desc'] = 'Here, you can upload a custom image (.ico or .png format) that the browser will show as the favicon of your Moodle website. If no custom favicon is uploaded, a standard Moodle favicon will be used.';
 // ... Section: Background images.
 $string['backgroundimagesheading'] = 'General background images';
-// ... Section: Login page background images.
-$string['loginbackgroundimagesheading'] = 'Login page background images';
-$string['loginbackgroundimage'] = 'Login page background images';
-$string['loginbackgroundimage_desc'] = 'The images to display as a background of the login page. One of these images will be picked randomly and shown when the user visits the login page.';
-$string['loginbackgroundimagetextsetting'] = 'Display text for login background images';
-$string['loginbackgroundimagetextsetting_desc'] = 'With this optional setting you can add text, e.g. a copyright notice to your uploaded background images. This text will appear on top of the page footer on the login page. However, for screen real estate reasons, it is only shown on larger screen sizes.<br/>
-Each line consists of the file identifier (the file name), the text that should be displayed and the text color, separated by a pipe character. Each declaration needs to be written in a new line. <br/>
-For example:<br/>
-background-image-1.jpg|Copyright: CC0|dark<br/>
-As text color, you can use the values "dark" or "light".<br />
-You can declare texts for a arbitrary amount of your uploaded login background images. The texts will be added only to those images that match their filename with the identifier declared in this setting.';
 // ... Section: Brand colors.
 $string['brandcolorsheading'] = 'Brand colors';
 // ... Section: Bootstrap colors.
@@ -104,6 +93,33 @@ $string['activityiconcolorcontentsetting_desc'] = 'The activity icon color for "
 // ... ... Setting: Activity icon color for 'Interface'.
 $string['activityiconcolorinterfacesetting'] = 'Activity icon color for "Interface"';
 $string['activityiconcolorinterfacesetting_desc'] = 'The activity icon color for "Interface"';
+
+// Settings: Login page tab.
+$string['loginpagetab'] = 'Login page';
+// ... Section: Login page background images.
+$string['loginbackgroundimagesheading'] = 'Login page background images';
+// ... ... Setting: Login page background image.
+$string['loginbackgroundimage'] = 'Login page background images';
+$string['loginbackgroundimage_desc'] = 'The images to display as a background of the login page. One of these images will be picked randomly and shown when the user visits the login page. Please make sure not to use non-ASCII-characters in the filename if you want to display text for login background images.';
+// ... ... Setting: Login page background image text.
+$string['loginbackgroundimagetextsetting'] = 'Display text for login background images';
+$string['loginbackgroundimagetextsetting_desc'] = 'With this optional setting you can add text, e.g. a copyright notice to your uploaded background images. This text will appear on top of the page footer on the login page. However, for screen real estate reasons, it is only shown on larger screen sizes.<br/>
+Each line consists of the file identifier (the file name), the text that should be displayed and the text color, separated by a pipe character. Each declaration needs to be written in a new line. <br/>
+For example:<br/>
+background-image-1.jpg|Copyright: CC0|dark<br/>
+As text color, you can use the values "dark" or "light".<br />
+You can declare texts for an arbitrary amount of your uploaded login background images. The texts will be added only to those images that match their filename with the identifier declared in this setting.';
+// ... Section: Login form.
+$string['loginformheading'] = 'Login form';
+// ... ... Setting: login form position.
+$string['loginformpositionsetting'] = 'Login form position';
+$string['loginformpositionsetting_desc'] = 'With this setting, you can optimize the login form to fit to a greater variety of background images. By default, the login form is displayed centered on the login page. Alternatively, you can move it to the left or to the right of the login page to let other parts of the background image shine through. Of course, you can also change this setting if no background images are uploaded at all.';
+$string['loginformpositionsetting_center'] = 'Centered';
+$string['loginformpositionsetting_left'] = 'Left-aligned';
+$string['loginformpositionsetting_right'] = 'Right-aligned';
+// ... ... Setting: login form transparency.
+$string['loginformtransparencysetting'] = 'Login form transparency';
+$string['loginformtransparencysetting_desc'] = 'With this setting, you can make the login form slightly transparent to let the background image shine through even more.';
 
 // Settings: Course tab.
 $string['coursetab'] = 'Course';
@@ -486,6 +502,42 @@ $string['flavourspreviewflavour'] = 'Preview flavour';
 $string['flavourspreviewblindtext'] = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Nunc id cursus metus aliquam eleifend mi in nulla. Felis imperdiet proin fermentum leo vel orci porta. Sed nisi lacus sed viverra tellus in hac habitasse. Vivamus arcu felis bibendum ut. Nisi porta lorem mollis aliquam ut porttitor. Odio euismod lacinia at quis risus sed vulputate odio. Sed felis eget velit aliquet sagittis id consectetur purus. Nec ullamcorper sit amet risus nullam eget. Pellentesque sit amet porttitor eget dolor. Cursus mattis molestie a iaculis at erat pellentesque.';
 $string['flavourstitle'] = 'Title';
 $string['flavourstitle_help'] = 'The flavour\'s title is just used internally to allow you to document a particular flavour in the list of flavours.';
+
+// Settings: Advertisement tiles.
+$string['tilestab'] = 'Advertisement tiles';
+// ... Section: Advertisement tiles general.
+$string['tilesgeneralheading'] = 'Advertisement tiles general';
+$string['tilecolumnssetting'] = 'Number of advertisement tile columns per row';
+$string['tilecolumnssetting_desc'] = 'Here, you define the number of columns per row in the presented grid of advertisement tiles. Please note that this number of columns applies to desktop / larger screens. On smaller screens and mobile screens, the advertisement tile columns are automatically wrapped.';
+$string['tilefrontpagepositionsetting'] = 'Position of the advertisement tiles on site home';
+$string['tilefrontpagepositionsetting_desc'] = 'Advertisement tiles are shown on site home only. With this setting, you control if the advertisement tiles are displayed before the site home content or after the site home content. If you want to show only the advertisement tiles on site home and nothing else, all other site home content can be removed by changing the <a href="{$a->url}">site home settings</a>.';
+$string['tilefrontpagepositionsetting_before'] = 'Before the site home content';
+$string['tilefrontpagepositionsetting_after'] = 'After the site home content';
+$string['tileheightsetting'] = 'Advertisement tiles height';
+$string['tileheightsetting_desc'] = 'With this setting, you control the height of the advertisement tiles. The configured height is the minimum height of each tile. If a tile\'s content is higher than this configured height, the whole row of tiles will be automatically made higher as needed.';
+// ... Section: Advertisement tiles.
+$string['tileheading'] = 'Advertisement tile no. {$a->no}';
+$string['tilebackgroundimagepositionsetting'] = 'Advertisement tile no. {$a->no} background image position';
+$string['tilebackgroundimagepositionsetting_desc'] = 'With this setting, you control the positioning of the background image within the advertisement tile no. {$a->no} container. The first value is the horizontal position, the second value is the vertical position.';
+$string['tilebackgroundimagesetting'] = 'Advertisement tile no. {$a->no} background image';
+$string['tilebackgroundimagesetting_desc'] = 'Here, you can upload an image file which will be shown as background image behind the content of the advertisement tile no. {$a->no}. Please make sure or check that the content is still readable on the background image. This is an optional setting, the advertisement tile will work even if you do not upload any background image.';
+$string['tilecontentsetting'] = 'Advertisement tile no. {$a->no} content';
+$string['tilecontentsetting_desc'] = 'Here, you enter the content which should be displayed in the advertisement tile no. {$a->no}. The content is displayed in the middle of the tile. This is an optional setting, the advertisement tile will be shown even if you do not set any content.';
+$string['tileenabledsetting'] = 'Enable advertisement tile no. {$a->no}';
+$string['tileenabledsetting_desc'] = 'With this setting, you can enable advertisement tile no. {$a->no}.';
+$string['tilelinksetting'] = 'Advertisement tile no. {$a->no} link URL';
+$string['tilelinksetting_desc'] = 'Here, you can set a (Moodle-internal or external) URL which will be offered as link button at the end of the advertisement tile no. {$a->no}. This is an optional setting, the advertisement tile will work even if you do not set any link URL.';
+$string['tilelinktitlefallback'] = 'Link';
+$string['tilelinktitlesetting'] = 'Advertisement tile no. {$a->no} link title';
+$string['tilelinktitlesetting_desc'] = 'Here, you can set a link title which is used as label of the link button as soon as you set a link URL in the advertisement tile no. {$a->no}. Please note that if you set a link URL but do not set a link title, the link button will just be labeled with \'Link\'.';
+$string['tilelinktargetsetting'] = 'Advertisement tile no. {$a->no} link target';
+$string['tilelinktargetsetting_desc'] = 'Here, you can set the link target which is set for the link button as soon as you set a link URL in the advertisement tile no. {$a->no}.';
+$string['tilelinktargetsetting_samewindow'] = 'Same window';
+$string['tilelinktargetsetting_newtab'] = 'New tab';
+$string['tileordersetting'] = 'Advertisement tile no. {$a->no} order position';
+$string['tileordersetting_desc'] = 'With this setting, you define the order position of the advertisement tile no. {$a->no}. By default, the advertisement tiles are ordered from top to bottom and left to right like you see them on this settings page here. However, you can decide to assign another order position with this setting. If you assign the same order position to two or more advertisement tiles, they will be ordered again according to the order on this settings page.';
+$string['tiletitlesetting'] = 'Advertisement tile no. {$a->no} title';
+$string['tiletitlesetting_desc'] = 'Here, you enter the title which should be displayed in the advertisement tile no. {$a->no}. This is an optional setting, the advertisement tile will be shown even if you do not set a title.';
 
 // Privacy API.
 $string['privacy:metadata'] = 'The Boost Union theme does not store any personal data about any user.';